[IMPORTANT]
Suggestion implementation. As per Anton's suggestion, I made the Card interface an enum.
Progress made on the game logic. I can play a very simple, basic version of the game, where the rules of Defusing & Exploding are implemented. That is, the player can only draw, as well as the Computer Player.